Update package.mustache
publish config is missing the dist folder which means that when using the package we are importing the TS files instead of the compiled js files and just reading the generated typings. This is an issue on all our apps using CRA, Vite, custom webpack etc. and I have seen a few issues raised for the same already. This is a non breaking change which would make peoples life easier, take it or leave it :)